+++ paleomix/common/argparse.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 import configargparse
 
+import paleomix
 
 SUPPRESS = configargparse.SUPPRESS
 
@@ -9,6 +10,16 @@ class ArgumentParser(configargparse.ArgumentParser):
     """Supports keys with underscores instead of dashes, for backwards compatibility
     with old paleomix config files, provided that these do not use per-host setions.
     """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        self.add_argument(
+            "-v",
+            "--version",
+            action="version",
+            version="%(prog)s v" + paleomix.__version__,
+        )
 
     def get_possible_config_keys(self, *args, **kwargs):
         keys = super().get_possible_config_keys(*args, **kwargs)