+++ paleomix/pipelines/zonkey/build_db.py
@@ -19,7 +19,6 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-import argparse
 import datetime
 import os
 import sys
@@ -28,6 +27,7 @@ import pysam
 
 from paleomix.common.sequences import NT_CODES
 
+import paleomix.common.argparse as argparse
 import paleomix.common.fileutils as fileutils
 import paleomix.pipelines.zonkey.common as common