+++ paleomix/pipelines/zonkey/build_mito.py
@@ -19,13 +19,13 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-import argparse
 import os
 import logging
 import sys
 
 import pysam
 
+import paleomix.common.argparse as argparse
 import paleomix.pipelines.zonkey.database as database
 
 from paleomix.common.formats.fasta import FASTA