+++ paleomix/tools/validate_fastq.py
@@ -21,14 +21,14 @@
 # SOFTWARE.
 #
 import sys
-import argparse
 import json
 
+from paleomix.common.argparse import ArgumentParser
 from paleomix.common.formats.fastq import FASTQ, FASTQualities
 
 
 def parse_args(argv):
-    parser = argparse.ArgumentParser("paleomix :validate_fastq")
+    parser = ArgumentParser("paleomix :validate_fastq")
     parser.add_argument("files", nargs="+")
     parser.add_argument("--collapsed", action="store_true")
     parser.add_argument("--no-empty", action="store_true")